created OKDUtils class
